requirements.txt
streamlit==1.35.0
pandas
openai==1.30.4
python-dotenv
